refactor(projections): give the projection API its own router
`make preflight` fails the product file line-count ratchet on the spike: the
projection routes and their response models pushed `routers/users.py` from its
2076-line baseline to 2152. The ratchet offers split-or-raise, and split is the
honest answer — the projection surface is its own product rather than part of
the user-profile router, and the planned rating route would only push it further.

- `routers/projections.py` takes the four routes and the three response models.
  Schema names are unchanged, so the generated app-client contracts and the Swift
  and TypeScript DTOs are unaffected; only the module that declares them moves.
- `users.py` returns to exactly 2076 lines, its existing baseline, so no bump.
- The three genuinely additive growths keep their files but raise their baselines
  with justifications: `storage.py` (+10, one upload helper following the
  `upload_app_logo` shape), `DesktopHomeView.swift` (+4, one switch case) and
  `SidebarView.swift` (+3, one navigation entry).
- Adds the desktop changelog fragment the spike's desktop changes require.

Verified against the real base. `make preflight` defaults to `origin/main`, which
is this fork's stale main — that reports a 1193-file diff of upstream commits and
demands invariant citations for files this branch never touched. Against
`upstream/main` (3bdf9af7d4), the true 17-file diff:

  python3 .github/scripts/pr_preflight.py --lane local --base upstream/main
  -> product-file-line-count-ratchet PASS (was FAIL on four files)
  -> product-invariants PASS — the diff touches no locked invariant path
  -> diff-hygiene, architecture-guardrails, failure-class-protocol PASS
  -> desktop-changelog-entry compares committed state, so it passes once this
     commit exists

  .venv/bin/python -c 'from routers import projections' -> the four paths are
  unchanged: /v1/users/projections/test, /v1/users/projections,
  /v1/users/projections/{projection_id}, /v1/projection-images/{id}.png
