+++ .github/scripts/product_file_line_count_ratchet_baseline/backend-utils.json
@@ -6,7 +6,7 @@
     "backend/utils/memory/canonical_memory_adapter.py": 2012,
     "backend/utils/memory/legacy_backfill.py": 1723,
     "backend/utils/memory_ingestion/pipeline.py": 1711,
-    "backend/utils/other/storage.py": 1584,
+    "backend/utils/other/storage.py": 1594,
     "backend/utils/retrieval/tools/calendar_tools.py": 1513,
     "backend/utils/sync/pipeline.py": 2491
   },
@@ -15,7 +15,7 @@
     "backend/utils/conversations/process_conversation.py": "Conversation finalization owns the transcript-to-memory transaction across canonical and legacy paths, so it supplies the shared dev-only allowlisted capture observer with the accepted-write result; cassette bounds, redaction, and export remain in testing.parity_pack_v0.live_capture.",
     "backend/utils/memory/canonical_consolidation.py": "Candidate selection, bounded retry leases, atomic promotion planning, and batch-level duplicate-supersede rejection remain one canonical long-term admission authority; extracting them would permit competing admission state machines (FC-split-mutation-authority).",
     "backend/utils/memory/canonical_memory_adapter.py": "Canonical conversation replacement coordinates source-generation fencing, bounded lineage discovery, and atomic replacement application; splitting this state machine would duplicate lifecycle authority (FC-split-mutation-authority).",
-    "backend/utils/other/storage.py": "delete_app_logo now requires an exact app-logo bucket prefix (startswith) so a foreign URL embedding the prefix later cannot delete an unrelated object on this deletion path (David review on #9873).",
+    "backend/utils/other/storage.py": "delete_app_logo now requires an exact app-logo bucket prefix (startswith) so a foreign URL embedding the prefix later cannot delete an unrelated object on this deletion path (David review on #9873); upload_projection_image follows the existing upload_app_logo shape so generated projection images use the same uploader rather than a second storage path (+10 lines).",
     "backend/utils/sync/pipeline.py": "Plan-aware soft-cap evaluation and its default-tier fallback remain in a small helper beside Sync's idempotent fresh-speech metering rather than overgrowing the durable worker coordinator."
   },
   "threshold": 1500
