+++ .github/scripts/product_file_line_count_ratchet_baseline/desktop-swift-mainwindow.json
@@ -1,25 +1,25 @@
 {
   "files": {
     "desktop/macos/Desktop/Sources/MainWindow/Components/ChatBubble.swift": 1869,
-    "desktop/macos/Desktop/Sources/MainWindow/DesktopHomeView.swift": 1634,
+    "desktop/macos/Desktop/Sources/MainWindow/DesktopHomeView.swift": 1638,
     "desktop/macos/Desktop/Sources/MainWindow/Pages/AppsPage.swift": 3626,
     "desktop/macos/Desktop/Sources/MainWindow/Pages/DashboardPage.swift": 4459,
     "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoriesPage.swift": 3530,
     "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoryGraph/CanonicalMemoryAtlasView.swift": 4427,
     "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoryGraph/MemoryAtlasForceLayout.swift": 1646,
     "desktop/macos/Desktop/Sources/MainWindow/Pages/TasksPage.swift": 5718,
-    "desktop/macos/Desktop/Sources/MainWindow/SidebarView.swift": 1690
+    "desktop/macos/Desktop/Sources/MainWindow/SidebarView.swift": 1693
   },
   "raise_justifications": {
     "desktop/macos/Desktop/Sources/MainWindow/Components/ChatBubble.swift": "Final-assistant Omi mark ownership and canonical expandable tool-call presentation are row-level rendering contracts that belong with bubble equality and layout. The mark remains an overlay so it cannot shift bubbles horizontally, while its owner row reserves the mark's height to prevent an empty streaming reply from clipping at the transcript live edge.",
-    "desktop/macos/Desktop/Sources/MainWindow/DesktopHomeView.swift": "Canonical Memory Atlas routing, unified Memory navigation, the cohort-not-yet-known guard, and leave handling coexist in the home content owner; the combined SwiftUI expression is split into independently type-checkable helper properties, with a further component split tracked separately.",
+    "desktop/macos/Desktop/Sources/MainWindow/DesktopHomeView.swift": "Retains the home routing and leave-handling owner while filling previously unused page slot 11 with Projections; the +4 lines are one switch case and its title mapping.",
     "desktop/macos/Desktop/Sources/MainWindow/Pages/AppsPage.swift": "Apps search now owns an explicit unknown/loading/completed/failed presentation state so the UI never flashes a false empty result before filtering completes.",
     "desktop/macos/Desktop/Sources/MainWindow/Pages/DashboardPage.swift": "Home ask bar now auto-grows as a multi-line TextField so long messages wrap instead of scrolling off-screen; the Home prompt rail also receives the ask bar's outer trailing edge so the custom rail aligns flush with it; a component split is tracked follow-up.",
     "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoriesPage.swift": "Memory search waits for its active projection, while canonical lifecycle capability, the Brain Map evidence inspector, and the split list/detail presentation remain owned by the Memories page; local-only identity resolution and its guarded delete/undo lifecycle remain with its optimistic mutation owner rather than exposing coordination state across view-model boundaries.",
     "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoryGraph/CanonicalMemoryAtlasView.swift": "The Canonical Memory Atlas owns progressive disclosure, island territories and captions, selection-trail Escape behavior, in-place evidence inspection, and related render policies. Geometry and territory labels are extracted where possible; a further surface split remains tracked separately.",
-    "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoryGraph/MemoryAtlasForceLayout.swift": "Relatedness layout, Louvain neighbourhood detection, and the group-separation and anchor-repulsion forces the island map is built on; no baseline entry existed because the file crossed the threshold in this change. It is already separate from the atlas view — a further split of detection from simulation is tracked follow-up.",
+    "desktop/macos/Desktop/Sources/MainWindow/Pages/MemoryGraph/MemoryAtlasForceLayout.swift": "Relatedness layout, Louvain neighbourhood detection, and the group-separation and anchor-repulsion forces the island map is built on; no baseline entry existed because the file crossed the threshold in this change. It is already separate from the atlas view \u2014 a further split of detection from simulation is tracked follow-up.",
     "desktop/macos/Desktop/Sources/MainWindow/Pages/TasksPage.swift": "Suggested-task loading feedback is now overlaid outside the list flow so an initial fetch cannot shift the task rows; a component split is tracked follow-up.",
-    "desktop/macos/Desktop/Sources/MainWindow/SidebarView.swift": "SB redesign: adds the hover-expand AppNavRail (icons→labels) and the merged/simplified navigation (Memory, Insights) driving selectedIndex; a component split is tracked follow-up."
+    "desktop/macos/Desktop/Sources/MainWindow/SidebarView.swift": "Adds the Projections entry to the navigation rail; +3 lines is one item in the list that owns sidebar identity and its selectedIndex mapping."
   },
   "threshold": 1500
 }
