+++ desktop/macos/changelog/unreleased/20260728-projections-page.json
@@ -1,0 +1,3 @@
+{
+  "change": "Added a Projections page that shows a generated image carrying one imperative drawn from your recent conversations"
+}
